# swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a Thermy backend API.",
        "title": "Thermy API",
        "contact": {
            "name": "API Support",
            "email": "andreysapozhkov535@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://mit-license.org/"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/elements": {
            "post": {
                "description": "add new elements of structural models",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "educator"
                ],
                "summary": "Add new elements of structural models.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Structural model element names",
                        "name": "modelElementNames",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.ModelElementNamesDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.ModelElementsIdDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/admin/layers": {
            "post": {
                "description": "add new text markup layer",
                "tags": [
                    "educator"
                ],
                "summary": "Add new text markup layer.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/admin/models": {
            "post": {
                "description": "add new structural models",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "educator"
                ],
                "summary": "Add new structural models.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Structural model names",
                        "name": "modelNames",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.ModelNamesDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.ModelsIdDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/educator/elements": {
            "post": {
                "description": "add new elements of structural models",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "educator"
                ],
                "summary": "Add new elements of structural models.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Structural model element names",
                        "name": "modelElementNames",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.ModelElementNamesDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.ModelElementsIdDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/educator/layers": {
            "post": {
                "description": "add new text markup layer",
                "tags": [
                    "educator"
                ],
                "summary": "Add new text markup layer.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/educator/models": {
            "post": {
                "description": "add new structural models",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "educator"
                ],
                "summary": "Add new structural models.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User authentication token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Structural model names",
                        "name": "modelNames",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.ModelNamesDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.ModelsIdDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.ModelElementNamesDTO": {
            "type": "object",
            "properties": {
                "model_element_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.ModelElementsIdDTO": {
            "type": "object",
            "properties": {
                "model_elements_id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.ModelNamesDTO": {
            "type": "object",
            "properties": {
                "model_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_Inspirate789_Thermy-backend_internal_domain_interfaces.ModelsIdDTO": {
            "type": "object",
            "properties": {
                "models_id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    }
}